requirements.txt
streamlit>=1.24.0
groq>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
requests>=2.31.0
googlemaps>=4.10.0 